django/requirements.txt
# /Users/yukiueta/business/development/anymore_meter/django/requirements.txt
Django==4.2.14
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.8.0
django-storages==1.14.3
django-simple-history==3.7.0
django-rest-passwordreset==1.3.0
djoser==2.1.0
celery==5.4.0
django-celery-beat
redis==5.0.1
boto3==1.35.10
mysqlclient==2.2.6
PyMySQL==1.0.2
python-decouple==3.8
python-dotenv==1.0.1
paho-mqtt
pycryptodome